move cellvit loader inside its dir
